chore: permutation builder refactor
