+++ bberg/src/lib.rs
@@ -3,6 +3,7 @@ mod circuit_builder;
 mod composer_builder;
 mod file_writer;
 mod flavor_builder;
+pub mod permutation_builder;
 mod prover_builder;
 mod relation_builder;
 mod utils;
